First draft of ADR on related pipelines
Part of the support for OLM operators within AppStudio is ensuring that
component relationships can be maintained. If there are components that
are dependent on each other, then a change in one image should be able
to precipitate a change in another image to ensure that the initial
change can be fully tested.

Signed-off-by: arewm <arewm@users.noreply.github.com>