Update BuildPipelineSelector example to new syntax
The BuildPipelineSelector CRD includes a pipelineRef field defined as a
Tekton PipelineRef. Previously, build-service used the Tekton v1beta1
APIs. The v1beta1.PielineRef [1] object supports specifying pipelines in
two different ways:

1) name + bundle:

    pipelineRef:
      name: mypipeline
      bundle: docker.io/myrepo/mycatalog:v1.0

2) resolver + params:

    pipelineRef:
      resolver: bundles
      params:
      - name: kind
        value: pipeline
      - name: name
        value: mypipeline
      - name: bundle
        value: docker.io/myrepo/mycatalog:v1.0

Build-service used to support only the first style. While migrating to
the Tekton v1 APIs, we added support for the second style.

For now, we maintain backwards compatibility with the first style, but
are planning to drop it soon. The v1.PipelineRef [2] object supports
only the second style.

Update the BuildPipelineSelector example to use the second style.

[1]: https://tekton.dev/docs/pipelines/pipeline-api/#tekton.dev/v1beta1.PipelineRef
[2]: https://tekton.dev/docs/pipelines/pipeline-api/#tekton.dev/v1.PipelineRef

Signed-off-by: Adam Cmiel <acmiel@redhat.com>