Update goals and constraints
Recently, more people are coming around the project to learn more about
it, how they might use it, and how they might contribute. The aim here
is to try to give them the most helpful high level introduction possible
before they get to the specifics further down in the doc and further out
in the repo.