Change architecture goal from recording in git to provenance
Semantic reproducibility should still be an architecture goal, but not
all parameters will be reasonable to commit to the git repository.
Instead of making this the goal, we should change the goal slightly to
require the information to be represented in the provenance and then
present in git where possible. This enables some workflows that leverage
tekton but not PAC as still within-goal.

Signed-off-by: arewm <arewm@users.noreply.github.com>