Generate API reference using Make
This was helpful to generate the API docs locally so I can preview and
run the link checker to check for broken links.

Adds the `tools/go.mod` so the version of `crd-ref-docs` can be pinned,
and the tool can be run using `go run`.

In essence one can do `make api-docs` to generate the Markdown files in
`content/ref`. There is a small issue with the
`service-provider-integration-operator`, it needs an additional
`go mod tidy` if golang version is newer than 1.20. To fix that run
`make api-docs` once to clone the repository, then
`cd crd-temp/service-provider && go mod tidy` and rerun `make api-docs`.