ADR 45: Guidelines for developing tasks
Presenting broad guidelines for where to contribute different types of
tasks as well as patterns that should be generally followed in the tasks
in the locations.

On the 2024-11-05 Konflux community architecture call, we discussed
whether there were any guidelines for where to contribute tasks to. The
result of that call was that we need to create an ADR outlining the
recommendations so that we can also enable more targeted conversation.

The recording for the call can be found on YouTube:
https://www.youtube.com/watch?v=uJbNdceDrTg

Signed-off-by: arewm <arewm@users.noreply.github.com>