feat: enable context efficient agent analysis
This repository is intended to be a record for how the various core and
add on services are designed to work and work with each other. When
designing or refining features, this type of information can be
instrumental when using ai assistants (i.e. claude).

This change creates a framework for lazy loading context into ai
assistants so that we can maximize the context available for analysis of
the features themselves.

Signed-off-by: arewm <arewm@users.noreply.github.com>