ADR for MintMaker log persistence
This commit adds an ADR regrading MintMaker log persistence strategy.

We will start integrating MintMaker with Konflux UI, and one of the
planned features is to show the logs of pipelineruns, so users can
check for themselves how things went.

For that to work, we will need to persist the log files. This ADR
discusses a few strategies for that.